@@ -1,16 +1,28 @@
 {
   "name": "rentdynamics",
-  "version": "0.6.2",
+  "version": "1.0.0",
   "description": "Package to help facilitate communicating with the Rent Dynamics API",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "types": "dist/esm/index.d.ts",
+  "type": "module",
+  "exports": {
+    "types": "./dist/esm/index.d.ts",
+    "node": "./dist/esm/index.js",
+    "require": "./dist/cjs/index.js",
+    "default": "./dist/esm/index.js"
+  },
   "scripts": {
-    "build": "npm run tsc && npm run webpack",
+    "build": "npm run build:esm && npm run build:cjs && npm run webpack",
+    "build:esm": "tsc -p tsconfig.esm.json",
+    "build:cjs": "tsc -p tsconfig.cjs.json",
     "nsp": "nsp check",
     "prepush": "npm run test",
     "test": "jest",
-    "tsc": "tsc",
-    "webpack": "webpack"
+    "webpack": "webpack",
+    "format": "prettier --write .",
+    "format:check": "prettier --check .",
+    "lint": "eslint lib"
   },
   "repository": {
     "type": "git",
@@ -30,22 +42,23 @@
   },
   "homepage": "https://github.com/RentDynamics/rentdynamics-js#readme",
   "devDependencies": {
-    "@types/chance": "^1.0.1",
-    "@types/jest": "^23.3.5",
-    "@types/jssha": "0.0.29",
-    "@types/node": "^10.12.18",
-    "chance": "^1.0.16",
+    "@eslint/js": "9.25.0",
+    "@types/chance": "1.1.6",
+    "@types/jest": "29.5.14",
+    "@types/node": "18.19.86",
+    "chance": "1.1.12",
     "codecov": "3.0.0",
+    "eslint": "9.25.0",
+    "globals": "16.0.0",
     "husky": "0.14.3",
-    "jest": "^23.6.0",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
     "nsp": "3.2.1",
-    "ts-jest": "^23.10.4",
-    "typescript": "^3.1.3",
-    "webpack": "^4.21.0",
-    "webpack-cli": "^3.1.2"
-  },
-  "dependencies": {
-    "cross-fetch": "3.0.0",
-    "jssha": "^2.3.1"
+    "prettier": "2.3.1",
+    "ts-jest": "29.3.2",
+    "typescript": "4.9.5",
+    "typescript-eslint": "8.30.1",
+    "webpack": "5.99.6",
+    "webpack-cli": "5.1.4"
   }
 }
