@@ -1,260 +1,287 @@
-import jsSHA from 'jssha';
+/** Payload is known as the "body" of a Request. The payload must be JSON serializable. */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type Payload = any;
 
-declare var global: any;
-
+/**
+ * Client is a convenience class for interacting with the Rent Dynamics services.
+ *
+ * Note if this class does not suit your needs, it may be wise to roll your own implementation using
+ * the {@linkcode ClientHelpers} class.
+ */
 export class Client {
+  /**
+   * authToken if defined, authToken is used to make authenticated requests. Note authToken is not
+   * typically get/set manually, but instead is managed through the login/logout methods.
+   */
+  public authToken: string | undefined;
+  private readonly helpers: ClientHelpers;
 
-    private helpers: ClientHelpers;
-    private options: ClientOptions;
-    public _fetch: any;
+  constructor(options: ClientOptions) {
+    this.helpers = new ClientHelpers(options);
+  }
 
-    constructor(options: ClientOptions) {
-        this.options = options;
-        this.helpers = new ClientHelpers(options);
-        if (typeof global !== "undefined" && global && global.fetch) {
-          this._fetch = global.fetch.bind(global);
-        } else {
-          let crossFetch = require('cross-fetch');
-          this._fetch = crossFetch;
-        }
+  /**
+   * get wraps a request library to work with the Rent Dynamics API.
+   * @param endpoint the path following the baseUrl.
+   * @example get('/foo');
+   */
+  public async get(endpoint: string) {
+    const fullUrl = this.helpers.baseUrl + endpoint;
+    const headers = await this.helpers.getHeaders(endpoint, undefined, this.authToken);
+    return fetch(fullUrl.replace(/\|/g, '%7C'), {
+      method: 'GET',
+      headers: {
+        ...headers,
+        'Content-Type': 'application/json'
+      }
+    });
+  }
+
+  /**
+   * put wraps a request library to work with the Rent Dynamics API.
+   * @param endpoint the path following the baseUrl.
+   * @param payload a JSON serializable object.
+   * @example put('/foo', { bar: 1 });
+   */
+  public async put(endpoint: string, payload: Payload) {
+    const fullUrl = this.helpers.baseUrl + endpoint;
+    const headers = await this.helpers.getHeaders(endpoint, payload, this.authToken);
+    return fetch(fullUrl, {
+      method: 'PUT',
+      headers: {
+        ...headers,
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(payload)
+    });
+  }
+
+  /**
+   * post wraps a request library to work with the Rent Dynamics API.
+   * @param endpoint the path following the baseUrl.
+   * @param payload a JSON serializable object.
+   * @example post('/foo', { bar: 1 });
+   */
+  public async post(endpoint: string, payload: Payload) {
+    const fullUrl = this.helpers.baseUrl + endpoint;
+    const headers = await this.helpers.getHeaders(endpoint, payload, this.authToken);
+    return fetch(fullUrl, {
+      method: 'POST',
+      headers: {
+        ...headers,
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(payload)
+    });
+  }
+
+  /**
+   * delete wraps a request library to work with the Rent Dynamics API.
+   * @param endpoint the path following the baseUrl.
+   * @example delete('/foo/1');
+   */
+  public async delete(endpoint: string) {
+    const fullUrl = this.helpers.baseUrl + endpoint;
+    const headers = await this.helpers.getHeaders(endpoint, undefined, this.authToken);
+    return fetch(fullUrl, {
+      method: 'DELETE',
+      headers: {
+        ...headers,
+        'Content-Type': 'application/json'
+      }
+    });
+  }
+
+  /**
+   * login enables an instance of {@linkcode Client} to make authenticated requests to the Rent
+   * Dynamics API.
+   */
+  public async login(username: string, password: string) {
+    const _password = await this.helpers.encryptPassword(password);
+    const endpoint = '/auth/login';
+    const result = await this.post(endpoint, { username, password: _password });
+    if (!result.ok) {
+      return result;
     }
+    const { token } = await result.json();
+    this.authToken = token;
+    return result;
+  }
 
-    public get(endpoint: string, parameters: object = {}): Promise<any> {
-        let queryParams = this.helpers.stringifyParameters(parameters);
-        let options: RequestInit = {};
-        options.method = 'GET';
-        options.headers = this.helpers.getHeaders(endpoint + queryParams);
-        let fullUrl = this.helpers.getBaseUrl() + endpoint + queryParams;
-        return this._fetch(fullUrl.replace(/\|/g, '%7C'), options).then((result: Response) => {
-          return result.ok ? result.clone().json().catch((err) =>
-              result.clone().text()
-          ) : result;
+  /** logout invalidates the users session generated by {@linkcode login}. */
+  public async logout() {
+    const endpoint = '/auth/logout';
+    const result = await this.post(endpoint, { authToken: this.authToken });
+    if (!result.ok) {
+      return result;
+    }
+    this.authToken = undefined;
+    return result;
+  }
+}
+
+/** BASE_URL is a collection of base urls for each dev/prod Rent Dynamics service. */
+export enum BASE_URL {
+  DEV_RD = 'https://api.rentdynamics.dev',
+  PROD_RD = 'https://api.rentdynamics.com',
+  DEV_RP = 'https://api-dev.rentplus.com',
+  PROD_RP = 'https://api.rentplus.com'
+}
+
+export type RdEncoder = Pick<TextEncoder, 'encode'>;
+
+export type RdCryptographer = Pick<SubtleCrypto, 'importKey'> &
+  Pick<SubtleCrypto, 'sign'> &
+  Pick<SubtleCrypto, 'digest'>;
+
+/** ClientOptions is consumed and updated by {@linkcode ClientHelpers}. */
+export class ClientOptions {
+  /** apiKey if defined apiKey is used to calculate auth headers. */
+  public apiKey: string | undefined;
+  /** apiSecretKey if defined apiSecretKey is used to calculate auth headers.  */
+  public apiSecretKey: string | undefined;
+  /**
+   * baseUrl is the base request url. The default is the development rentdynamics api. A custom
+   * string may be provided beyond the {@linkcode BASE_URL} options.
+   */
+  public baseUrl: BASE_URL | string = BASE_URL.DEV_RD;
+
+  /**
+   * getEncoder is used to encode text. The encoder can be overridden as needed. For example in a
+   * node environment.
+   * @example
+   * const options = new ClientOptions();
+   * options.getEncoder = async () => new (await import('util')).TextEncoder();
+   */
+  public getEncoder = async (): Promise<RdEncoder> => new TextEncoder();
+
+  /**
+   * getCryptographer is used for cryptography. The cryptographer can be overridden as needed. For
+   * example in a node environment.
+   * @example
+   * const options = new ClientOptions();
+   * options.getCryptographer = async () => (await import('crypto')).subtle;
+   */
+  public getCryptographer = async (): Promise<RdCryptographer> => crypto.subtle;
+}
+
+/**
+ * ClientHelpers is a collection of utilities consumed by {@linkcode Client}. ClientHelpers can be
+ * used to calculate headers in case a consumer wants to build their own API client.
+ */
+export class ClientHelpers {
+  private options: ClientOptions;
+
+  constructor(options: ClientOptions) {
+    this.options = options;
+  }
+
+  /**
+   * baseUrl is the base url used throughout the {@linkcode ClientHelpers} instance. It is initially
+   * configured through {@linkcode ClientOptions}.
+   */
+  get baseUrl() {
+    return this.options.baseUrl;
+  }
+
+  /**
+   * getTimestamp is used to calculate the timestamp header. This method is not likely to be called
+   * on it's own. Instead, it is typically used to mock the current time.
+   */
+  public getTimestamp() {
+    return Date.now();
+  }
+
+  /**
+   * getHeaders creates headers for the given params. If an auth token is included, this method will
+   * generate an `Authorization` header.
+   */
+  public async getHeaders(
+    endpoint: string,
+    payload?: Payload | undefined,
+    authToken?: string | undefined
+  ) {
+    const headers: Record<string, string> = {};
+    if (this.options.apiKey && this.options.apiSecretKey) {
+      if (typeof payload !== 'undefined') {
+        payload = this.formatPayload(payload);
+        payload = JSON.stringify(payload);
+        if (['{}', '[]'].includes(payload)) {
+          payload = '';
+        }
+      }
+      const timestamp = this.getTimestamp();
+      const nonce = await this.getNonce(timestamp, endpoint, payload);
+      if (authToken) {
+        headers.Authorization = 'TOKEN ' + authToken;
+      }
+      headers['x-rd-api-key'] = this.options.apiKey;
+      headers['x-rd-api-nonce'] = nonce;
+      headers['x-rd-timestamp'] = timestamp.toString();
+      return headers;
+    }
+    return headers;
+  }
+
+  /** formatPayload formats the payload for nonce calculation. */
+  public formatPayload(payload: Payload): Payload {
+    let formattedPayload: Payload = {};
+    if (payload === undefined || payload === null) {
+      formattedPayload = null;
+    } else if (payload !== Object(payload)) {
+      formattedPayload = payload;
+    } else if (Array.isArray(payload)) {
+      formattedPayload = [];
+      payload.forEach((_, index) => {
+        formattedPayload[index] = this.formatPayload(payload[index]);
+      });
+    } else {
+      Object.keys(payload)
+        .sort()
+        .forEach(k => {
+          if (typeof payload[k] === 'object') {
+            formattedPayload[k] = this.formatPayload(payload[k]);
+          } else if (typeof payload[k] === 'string') {
+            formattedPayload[k] = payload[k].replace(/ /g, '');
+          } else {
+            formattedPayload[k] = payload[k];
+          }
         });
     }
+    return formattedPayload;
+  }
 
-    public put(endpoint: string, payload: object): Promise<any> {
-      let options: RequestInit = {};
-      options.method = 'PUT';
-      options.headers = this.helpers.getHeaders(endpoint, payload);
-      options.body = JSON.stringify(payload);
-      let fullUrl = this.helpers.getBaseUrl() + endpoint;
-      return this._fetch(fullUrl, options).then((result: Response) => {
-          return result.ok ? result.clone().json().catch((err) =>
-              result.clone().text()
-          ) : result;
-      });
-    }
+  /** getNonce calculates the nonce for the given params. */
+  public async getNonce(timestamp: number, url: string, payloadStr?: string): Promise<string> {
+    if (!this.options.apiSecretKey) return Promise.resolve('');
+    const encodedUrl = encodeURI(url)
+      .replace(/%7[Cc]/g, '|')
+      .replace(/%20/g, ' ');
+    const nonceStr =
+      typeof payloadStr !== 'undefined'
+        ? timestamp + encodedUrl + payloadStr
+        : timestamp + encodedUrl;
+    const cryptographer = await this.options.getCryptographer();
+    const encoder = await this.options.getEncoder();
+    const key = encoder.encode(this.options.apiSecretKey);
+    const data = encoder.encode(nonceStr);
+    const algorithm = { name: 'HMAC', hash: 'SHA-1' };
+    const hmac = await cryptographer.importKey('raw', key, algorithm, false, ['sign']);
+    const signed = await cryptographer.sign(algorithm.name, hmac, data);
+    return _hexDigest(signed);
+  }
 
-    public post(endpoint: string, payload: object): Promise<any> {
-      let options: RequestInit = {};
-      options.method = 'POST';
-      options.headers = this.helpers.getHeaders(endpoint, payload);
-      options.body = JSON.stringify(payload);
-      let fullUrl = this.helpers.getBaseUrl() + endpoint;
-      return this._fetch(fullUrl, options).then((result: Response) => {
-          return result.ok ? result.clone().json().catch((err) =>
-              result.clone().text()
-          ) : result;
-      });
-    }
-
-    public delete(endpoint: string): Promise<any> {
-      let options: RequestInit = {};
-      options.method = 'DELETE';
-      options.headers = this.helpers.getHeaders(endpoint);
-      let fullUrl = this.helpers.getBaseUrl() + endpoint;
-      return this._fetch(fullUrl, options).then((result: Response) => {
-          return result.ok ? result.clone().json().catch((err) =>
-              result.clone().text()
-          ) : result;
-      });
-    }
-
-    public login(username: string, password: string): Promise<any> {
-      let shaObj = new jsSHA('SHA-1', 'TEXT');
-      shaObj.update(password);
-      password = shaObj.getHash('HEX');
-      let endpoint = '/auth/login';
-      return this.post(endpoint, { username: username, password: password }).then((result) => {
-        this.options.authToken = result.token;
-        return result;
-      });
-    }
-
-    public logout(): Promise<any> {
-      let endpoint = '/auth/logout';
-      return this.post(endpoint, {authToken: this.options.authToken}).then((res) => {
-        this.options.authToken = undefined;
-        return res;
-      });
-    }
-
+  /** encryptPassword encrypts the password for login. */
+  public async encryptPassword(password: string) {
+    const cryptographer = await this.options.getCryptographer();
+    const encoder = await this.options.getEncoder();
+    const encodedPassword = encoder.encode(password);
+    const digestedPassword = await cryptographer.digest('SHA-1', encodedPassword);
+    return _hexDigest(digestedPassword);
+  }
 }
 
-export class ClientOptions {
-    public apiKey?: string = undefined;
-    public apiSecretKey?: string = undefined;
-    public authToken?: string = undefined;
-    public development?: boolean = false;
-    public service?: string = undefined;
-    public developmentUrl?: string = undefined;
-    public baseUrl?: string = undefined;
-}
-
-export class ClientHelpers {
-    private options: ClientOptions;
-
-    constructor(options: ClientOptions) {
-        this.options = options;
-    }
-
-    public formatPayload(payload: any): any {
-        let formattedPayload:any = {};
-
-        if (typeof payload === undefined || payload === null) {
-            formattedPayload = null;
-        } else if (payload !== Object(payload)) {
-            formattedPayload = payload;
-        } else if (Array.isArray(payload)) {
-            formattedPayload = [];
-
-            for (let i = 0; i < payload.length; i++) {
-                formattedPayload[i] = this.formatPayload(payload[i]);
-            }
-        } else {
-            Object.keys(payload).sort().forEach( (k:any, v:any) => {
-                if (typeof(payload[k]) == 'object') {
-                    formattedPayload[k] = this.formatPayload(payload[k]);
-                } else if (typeof (payload[k]) == 'string') {
-                    formattedPayload[k] = payload[k].replace(/ /g, '');
-                } else {
-                    formattedPayload[k] = payload[k]
-                }
-            }, this);
-        }
-        return formattedPayload;
-    }
-
-    public getBaseUrl() {
-        if (this.options.development && this.options.developmentUrl) {
-            return this.options.developmentUrl;
-        } else if (this.options.development) {
-            return 'https://api-dev.rentdynamics.com';
-        }
-        if (this.options.baseUrl) {
-            return  this.options.baseUrl;
-        } else {
-            return 'https://api.rentdynamics.com';
-        }
-    }
-
-    public getHeaders(endpoint: string, payload?: Object) {
-        let headers: any = {};
-        if (this.options.apiKey && this.options.apiSecretKey) {
-          if (typeof payload !== "undefined") {
-                payload = this.formatPayload(payload);
-            }
-            let timestamp = Date.now();
-            let nonce = this.getNonce(timestamp, endpoint, JSON.stringify(payload));
-            if (this.options.authToken) {
-                headers['Authorization'] = 'TOKEN ' + this.options.authToken;
-            }
-            headers['x-rd-api-key'] = this.options.apiKey;
-            headers['x-rd-api-nonce'] = nonce;
-            headers['x-rd-timestamp'] = timestamp.toString();
-            headers['Content-Type'] = 'application/json';
-            return headers;
-        }
-        return headers;
-    }
-
-    public getNonce(timestamp: number, endpoint: string, payloadStr?: string) {
-        if (this.options.apiSecretKey) {
-            let nonceStr = timestamp + endpoint;
-
-            if (typeof payloadStr !== 'undefined') {
-                nonceStr += payloadStr;
-            }
-            let shaObj = new jsSHA('SHA-1', 'TEXT');
-            shaObj.setHMACKey(this.options.apiSecretKey, 'TEXT');
-            shaObj.update(nonceStr);
-
-            return shaObj.getHMAC('HEX');
-        } else {
-            return '';
-      }
-    }
-
-    public stringifyParameters(object: any): string {
-      let results = [];
-
-      if (object.filters && this.stringifyFilters(object.filters))
-        results.push(`filters=${this.stringifyFilters(object.filters)}`);
-
-      if (object.include && object.include.length)
-        results.push(`include=${object.include.join(',')}`);
-
-      if (object.exclude && object.exclude.length)
-        results.push(`exclude=${object.exclude.join(',')}`);
-
-      if (object.fields && object.fields.length)
-        results.push(`fields=${object.fields.join(',')}`);
-
-      if (object.orderBy)
-        results.push(`orderBy=${object.orderBy}`);
-
-      if (object.page)
-        results.push(`page=${object.page}`);
-
-      if (object.pageSize)
-        results.push(`pageSize=${object.pageSize}`);
-
-      if (object.distinct)
-        results.push(`distinct=true`);
-
-      return results.join('&') ? `?${results.join('&')}` : '';
-    }
-
-    private mapKeyValuePairs(key: string, value: any): string[] {
-      let results: string[] = [];
-
-      if (Array.isArray(value)) {
-        if (value.length)
-          results.push(`${key}__in=${value.join(',')}`)
-        return results;
-      }
-
-      if (value !== null && typeof value === 'object') {
-        for (var filterValKey in value) {
-          let result: string[];
-
-          result = this.mapKeyValuePairs(`${key}__${filterValKey}`, value[filterValKey]);
-
-          if (result && result.length)
-            results.push(result.join('|'));
-        }
-        return results;
-      }
-
-      if (value !== null && value !== undefined && value !== '') {
-        results.push(`${key}=${value}`);
-        return results;
-      }
-
-      return results;
-    }
-
-    private stringifyFilters(filter: {[key: string]: any}): string {
-      let results: string[] = [];
-
-      for (var key in filter) {
-        let result: string[] = this.mapKeyValuePairs(key, filter[key]);
-
-        if (result && result.length)
-          results.push(result.join('|'));
-      }
-
-      return results.join('|');
-    }
-
-}
+const _hexDigest = (buf: ArrayBuffer): string =>
+  Array.from(new Uint8Array(buf))
+    .map(b => b.toString(16).padStart(2, '0'))
+    .join('');
